+++ packages/at_client/CHANGELOG.md
@@ -1,4 +1,5 @@
 ## 3.0.77
+- fix: Fix the keys expiry job not being triggered
 - chore: deprecate NotificationParams.forText()
 ## 3.0.76
 - feat: Introduce mechanism to identify and delete expired keys